[feature/mainpage] fix : layout 코드 분리 및 등급표 아이콘 수정
